+++ app/src/main/java/cn/coegle18/smallsteps/entity/Account.kt
@@ -4,10 +4,10 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class Account(val accountType: Int,
+data class Account(val accountType: Long,
                    val name: String,
                    var visible: Boolean,
-                   var billNum: Int,
+                   var billNum: Long,
                    var balance: Double,
                    var remark: String) {
     @PrimaryKey(autoGenerate = true)
