+++ app/src/main/java/cn/coegle18/smallsteps/entity/AccountType.kt
@@ -2,7 +2,9 @@ package cn.coegle18.smallsteps.entity
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import cn.coegle18.smallsteps.*
+import cn.coegle18.smallsteps.BaseAccountType
+import cn.coegle18.smallsteps.MainAccountType
+import cn.coegle18.smallsteps.PrimaryAccountType
 
 @Entity
 data class AccountType(val finalType: Boolean,
@@ -10,13 +12,13 @@ data class AccountType(val finalType: Boolean,
                        val primaryAccountType: PrimaryAccountType,
                        val name: String,
                        val mainAccountType: MainAccountType,
-                       val parentId: Int,
+                       val parentId: Long,
                        val visible: Boolean,
-                       val custom:Boolean,
+                       val custom: Boolean,
                        val autoImport: Boolean,
                        val hint: String,
                        val icon: String
-                       ) {
+) {
     @PrimaryKey(autoGenerate = true)
     var accountTypeId: Long = 0
 }
